Refactored api file structure, api_ for wrappers
