+++ server/api/api_users_add_event.go
@@ -4,8 +4,6 @@
 package api
 
 import (
-	"github.com/pkg/errors"
-
 	"github.com/mattermost/mattermost-plugin-solar-lottery/server/utils/bot"
 )
 
@@ -30,20 +28,6 @@ func (api *api) AddEvent(mattermostUsernames string, event Event) error {
 	}
 
 	logger.Infof("%s added event %s to %s.",
-		api.MarkdownUser(api.actingUser), MarkdownEvent(event), api.MarkdownUsersWithSkills(api.users))
-	return nil
-}
-
-func (api *api) addEventToUsers(users UserMap, event Event, persist bool) error {
-	for _, user := range users {
-		user.AddEvent(event)
-
-		if persist {
-			_, err := api.storeUserWelcomeNew(user)
-			if err != nil {
-				return errors.WithMessagef(err, "failed to update user %s", api.MarkdownUser(user))
-			}
-		}
-	}
+		api.MarkdownUser(api.actingUser), api.MarkdownEvent(event), api.MarkdownUsersWithSkills(api.users))
 	return nil
 }
