+++ server/api/api_users_delete_event.go
